# Save state to DB after proposer boost