This PR optimizes the time-to-set proposer boost. It shreds off the wait time by ~50%

Based on the local setup, it takes ~100ms to set proposer boost, ~50ms of that is saving state and blocks to DB. We can move them after setting proposer boost.

```
# Before 
state_transition (30ms)
notify_new_payload (10ms)
save_block_to_forchoice (1ms)
save_state_block_to_db (40ms)
set_proposer_boost

# After
state_transition (30ms)
notify_new_payload (10ms)
save_block_to_forchoice (1ms)
set_proposer_boost
save_state_block_to_db (40ms)
```

Discovered in:
![](https://cdn.discordapp.com/attachments/774005284188323870/962919464108695652/on_block-rev1.png)